requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5,<1.24.0
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
timm
flake8
protobuf==3.20.1

#protobuf<4.21.3
#torchsummary
# Logging -------------------------------------
#tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
#pandas>=1.1.4
#seaborn>=0.11.0

# Export --------------------------------------
onnx==1.7.0 
onnxruntime==1.16.3
onnx-simplifier>=0.3.6  # ONNX simplifier

# Extras --------------------------------------
#ipython  # interactive notebook
#psutil  # system utilization
#thop  # FLOPs computation
#albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
